Comment the gem install command that makes pry available.
Pry is handy to debug tests.  You can run the test from the command line,

      $ ruby test.rb

and when it hits a "binding.pry" statement, it enters into the REPL.  This is
useful when you are building a new test; you can setup a baseline and then
interactively figure out the remaining lines to add to the test.