Fix broken breadcrumb in news and API docs.
Another surprising bug.  The link back to the index of news stories was
broken; clicking it kept you on the same page (the full news story).  I guess
this module was not used much.

While here, fix the same issue with the API doc breadcrumbs.

It looks like the simple survey has the same issue, but I'm not testing that
now so I'll leave for later.  Also, the PR archive might also have the issue.

To look for the problem files, run this from the shell:

	$ find . -name "*.tcl" | xargs grep ad_context_bar_ws_or_index | grep "\"\""