Create static directory for static assets.
Setting the foundation for reducing GET's.  Plan is to create a
shell script that puts separate JavaScript files together, minifies,
then changes file name to include the MD5 sum, and finally updates
the properties file to point to this generated file name.  This
will we can use expires = never for asset's in the static directory.

Having one directory where we know all these assets live simplifies the logic
a bit.