Change logic in displaying news index.
Previous logic only displayed the news story in the index if it was in the top
3 stories, had no comments, and was less than 300 characters long.

This change displays the news story if it meets those conditions and is
anywhere in the list of current news articles.

Also, when we do display the news article in the index, include the title.

Also, move the header tag into the news index.tcl, until I can find
(or build) a more general location.