Start cutting down reveal.js.
In this commit, don't do any layout with Javascript.  The zooming
and aspect ratio logic didn't work so well with someone resizing
their browser in lots of different ways.  Usually, the font ended
up way too small.  To be fair, that's not what it was designed for.

We will rely on pagination and responsive styling.